:art: Rename english template
